feature 合并sso分支代码到dev分支：（1）web平台支持SSO （2）支持普通用户和管理员两种页面 （3）新增配置项，改造定时任务获取TPS的开关 （4）新增大量普通用户所看页面

+++ gmq-web/src/main/java/com/alibaba/rocketmq/service/gmq/GMQSysResourceService.java
@@ -221,10 +221,10 @@ public class GMQSysResourceService extends AbstractService {
             String version = kvTable.getTable().get(BrokerConst.BROKER_VERSION);
 
             String[] inTpsValues = putTps.split(" ");
-            long in = inTpsValues != null && inTpsValues.length > 0 ? Long.parseLong(inTpsValues[0]) : 0L;
+            long in = inTpsValues != null && inTpsValues.length > 0 ? ((Double)Double.parseDouble(inTpsValues[0])).longValue() : 0L;
 
             String[] outTpsValues = getTransferedTps.split(" ");
-            long out = outTpsValues != null && outTpsValues.length > 0 ? Long.parseLong(outTpsValues[0]) : 0L;
+            long out = outTpsValues != null && outTpsValues.length > 0 ? ((Double)Double.parseDouble(outTpsValues[0])).longValue() : 0L;
 
             String msgPutTotalYesterdayMorning = kvTable.getTable().get(BrokerConst.PUT_TOTAL_YESTERDAY_MORNING);
             String msgPutTotalTodayMorning = kvTable.getTable().get(BrokerConst.PUT_TOTAL_TODAY_MORNING);
